Update net.wooga.atlas-paket to 0.8.0
Description
===========

Update the internal used plugin `net.wooga.atlas-paket` to `0.8.1`.
This new version changed the way it determines the default version value
for a `PaketPack` task. If not set directly it will first check for a
version value in the `paket.template` file and then, if this value is
empty, will use the project version. We want to enforce the use of the project
version for our builds so we set the version property of all tasks with
type `PaketPack` to current project version. This ensures that we don't
generate wrong package versions even if a `paket.template`
contains a `version` argument.

Changes
=======

![Update] `net.wooga.atlas-paket` to `0.8.1`
![Change] set version property on all tasks with type `PaketPack`