Fix error messages for rc builds (#25)
Description
===========

The `NebularRelease` plugin will provide slightly better error messages when using the official
cli tasks (final, candidate, snapshot, ...). Because of internal naming reasons it makes sense for us to use
`rc` instead of `candidate`. All other resources are named with the
pattern (final, rc and snapshot). I used a custom task with the name `rc` which depends on
`candidate` but this will fall through the error check in `NebularRelease`. So instead we
now change the cli tasklist on the fly. If we find the `rc` taskname in the cli tasklist we remove it
and add `candidate` instead.

Changes
=======

![ADD] alias from task `rc` to `candidate`
![REMOVE] custom task `rc`
![IMPROVE] error messages for `rc` builds