Fix branch push during release
Description
===========

The `BaseReleasePlugin` will automatically push the current branch when
the name != `HEAD`. This leads for us to some funny sideeffects on the
Jenkins machines. This behavior is not configurable at this time. What
this pull request does is override the `release` task action and provide
a more conservative implementaton.

Issues
------

While implementing I was wondering if we really want to stop every push
or should make this somehow configurable on our side.

Changes
=======

![IMPROVE] `release` task. Remove `push` instruction.