Fix version precedence check for snapshot (ci) builds
Description
===========

The new `semver 2.0` snapshot strategy enforces version precedence due
to a configuration error. This is a clear bug because snapshot builds
are designed to be outside the version order for `rc` or `final` builds.
So how can this bug show itself. The only valid way at the moment to
trigger this bug is to build a prerelease build (`rc`) and then a
snapshot version with the same base version. This can happen in
`release/2.x` branches. A snapshot build should have no direct
connections to any pre or final release.

**gradle output**
```
> Inferred version (2.0.0-branch.release.2.x.0) cannot be lower than nearest (2.0.0-rc.1). Required by selected strategy.
```

This patch simply configures the snapshot strategy not to enforce
version precedence (the old v1 did this as well).

Changes
=======

![FIX] semver `2.0` snapsphot version prcedence check