Fix branch name conversion
Description
===========

Git allows a few combinations of characters in branch names which are
not allowed in a semver version. The plugin already takes take by
replacing the special characters: `/_-` depending on the semver scheme.
But the regex parsing these characters are not checking for a series of
these special characters.
Branch names like these would fail to create a valid version:

`feature/-a_feature`
`fix/some_-random-bug`

I adjusted the regular expressions for both semver1 and semver2
strategies. The adjustments will result in a cleaner version generation
but would also allow the creation of weird branch names. Since these
names are valid git branch names I opted for a patch of the regex
instead of a branch name validator.

Semver1
-------

The strategy for the WDK semver1 versions has a slight drawback. When
creating it I decided to make the `.` character a special one and
replaced it with `Dot`. I forgot to capitalize the next word after the
`.` which becomes apperant only now.
A branchname like `fix/-_.bug` would yield a version like
`1.0.1-branchFixDotbug00009`. See the word _bug_ is not capitalized.
But a branchname like `fix/.-bug` would yield
`1.0.1-branchFixDotBug00009`. I don't want to change this now. We went
once changeing the version scheme and small things like these can have
unwanted repercussions.

Semver2
-------

Here the patch is straight forward. Replace a series of delimiter characters
including `.` followed by a word character with `.` + the word
character.

Changes
=======

* ![FIX] branch name conversion